Remove unnecessary notes from GERLSHMUD
As part of the split of GERLSHMUD into EGRE, EGRE_MUD and EGRE_MUD_1 I'm
removing files that don't pertain to the specific rules for a particular
MUD.

These notes all remain in the EGRE project and have been converted from
HTML to text. They were duplicated in the spit and can be removed.

A couple files are renamed but still have HTML content. It's easier to
rename in one step (to retain the git history) and then update the
contents in another step.