Rename all source files from GERLSHMUD to mud_
I want the module prefix to be short. I don't think this project will be
a dependency of other MUD projects since it's specific to EGRE_MUD; the
mud_ prefix shouldn't collide. (Let's be real, no is ever going to use
this so collisions are not a concern.)

We don't need all these files but the rename is easy.