Handlers -> Rules
Handlers "from graph of Erlang stream handlers" are now known as rules,
as in "Erlang graph rules engine".

The behaviour that rules implement is now "egre_rules", because each
rules module has multiple rules in it (i.e. as opposed to "egre_rule",
implying that each behaviour implementation is a single rule).