Update notes
I still thinking keeping a file of notes about the project in the
project is a good idea. I try and put a lot of "why" comments in the git
commit comments (such as this one), but I think a long-running note file
is a good idea for keeping track of the general thought processes that
went into the project.