All game config belongs to the MUD
The MUD must contain all of the descriptions for elements of the MUD.
Since building descriptions also relies on the MUD config, the
mud_util.erl module needs to be owned by the MUD as well.

Hence, mud_config.erl and mud_util.erl need to be used instead of
egre_config.erl and egre_util.erl.