More "handlers" -> "rules"
I don't remember all the details, but attacks create effects, using an
effect "prototype"; the prototype creates an effect object with rules
that are stored in the "child_rules" property. I was looking in the
"child_handlers" property for the child rules, and adding a "handlers"
property: this had to change to "child_rules" and "rules".

Also, the player_attack test generates enough data in four seconds; we
don't need to run it for 20 seconds.