Add a "hoverable" capability in the HTML logs
I want to allow anything to be hovered over in the logs so you can see
what it means; sometimes I haven't worked on this project in months and
I can't remember what the various fields mean.

The CSS is a bit weird:
It uses the tilde "sibling" operator: if you have an element with the
class "hoverable", and you hover over it, and it has a sibling with the
class "hoverLabel", then the hover label will become visible. Because of
the z-index, the hover label will float above the other elements on the
page, i.e. it will hover.

I wrapped this up in a "hoverSpan" function so that I could easily wrap
the "hoverable" element, and its hover label, in a span, so that they
were the only siblings. The tilde operator might only work with the
nearest sibling, but I don't feel like checking. Also, this allows me to
wrap an element that already has a CSS class.

Now you can hover over the "filter" checkboxes, the event type circles
(i.e. attempt or succeed), and the rules module that triggered the
event, if any.