Make "mud" a "library application"
By removing the 'mod' key in the .app.src file, the application because
a "library" application and doesn't start an application process or a
supervisor. (Actually, I don't know if there _is_ an application
process, or if it's just the supervisor.)

I want egre_mud_1 to have application env vars, but I don't want to have
a supervisor that starts up and does nothing.