Rebuild egre when it changes
This was an attempt to get erlang.mk to rebuild the egre dependency when
it changes. I'm not sure if it worked.

My original goal was to allow egre_mud_1 to only require egre_mud, and
then egre_mud would depend on egre; i.e. egre_mud_1 shouldn't know about
egre; however, that's dumb because egre_mud_1 consists mostly of rules
that conform to the egre spec. You can't start building a MUD on top of
egre_mud without knowing you're using egre.

It would still be nicer if you only had one dependency in egre_mud, but
I'm not going to cry myself to sleep over it.

Also, you can just cd in deps/egre and run make; it builds the
dependencies of egre the first time, but that's not the end of the
world.