Mostly naming fixes
I still had a lot of "gerlshmud" names kicking around. I wasn't referring
to the "mud" application for the corpse lifetime: egre_mud shouldn't
know anything about any particular MUD.

I think we increased the wait in the "player_remove" test to see if we
were giving up to early, but I don't think that was the problem. That
can probably go back down to a 100 ms wait.