Stop top level app
If we don't stop mud, and just stop egremud, then on subsequent test
runs, if we try and start mud, it's already started, and erlang,
apparently, doesn't look any further into the dependency graph.

i.e. Erlang will see that the mud app is started and not make sure that
egremud or egre are started.