Match HTML log pids in message with id
Events in the MUD can have pids in them. It's annoying to figure out
what the pids are; you can show the entire log of different messages and
see what ..._id fields match which pid fields, or you can look at the
"<source> <verb> <target>" field.

I changed the JavaScript to go through every log message and catalog all
the PID -> ID pairs. There's lots of duplicates, e.g. object, source,
target, parent, etc. but they all match. For example, in one log message
you might have:
  ░ parent: <0.0.1>
  ░ parent_id: player
Then, in another log message you might have:
  ░ target: <0.0.1>
  ░ target_id: player
The PID -> ID mapping is the same in both, even though the log field is
different. This is because all log fields use the same egre index to
find the ID for a PID when the log message is generated.

So here's an original message

<0.278.0> move from <0.264.0> to <0.274.0> limited to legs

... and here's the new message:

pants               giant        legs0
<0.278.0> move from <0.264.0> to <0.274.0> limited to legs

Note that the IDs are written above the original PIDS so that the
original message is clear, but the names are still easy to see. It's
important to be able to see the original message as-is.

I also added a "<no message>" message for logs that don't have a
message. So far this is only for debug log messages.

Dev Note:

I didn't realize that some events are logged as an array of strings, and
some events are logged as strings. I originally only knew about the
strings, so I had some elaborate code to regex match the pids and then
generate an array of non-PID text, interspersed with {pid: pid} objects.
While debugging this, I figured out that some events are arrays, so I
just went through each array element looking for pids. Once I did that,
I realized that pids never have spaces, so I could just split the string
events on spaces and then look for pids the same as I do for arrays. It
was a little painful to throw out my clever partitioning code, but now
it's much simpler.