Fix look bugs, find new look bugs
There was a problem with recursion when looking at a character: the body
part was responding to {..., describe, Owner, ...} and {..., describe,
Self, ...}; The "self" version is for the benefit of the body part's
children. A body part only needs to describe itself if it gets looked at
directly, or if its parent is looked at.

We also created another test to look directly at a body part, but we
haven't verified if it passes.

There's some commented out code that I need to get rid of.