Add player shout test
A player shouting is heard from one room over.
Test that:
- player in the room that shouted hears it
- player in the room that didn't shout hears it
- player in the next room (kitchen) hears it
- player two rooms away (bathroom) doesn't hear it

The login logic was getting repetitive in the tests, so I added a login
function.

I got tired of relying on the ?WAITnnn macros, and couldn't figure out
why I didn't just use a function, so I converted them to a function.

Added a wait loop with no predicate to empty the socket "buffers".

Add ability for players to shout, because that's on the list of things I
want the reference MUD to do. (i.e. to show how one might do it.)

Remove some cruft. Should be obvious why.