Duplicate WORLD_3 to make changes
I needed to make a change to WORLD_3 in order to fix an intermittent
error in the player_resource_wait code, but I didn't want to risk
breaking other tests. Victor Savkov suggested using world names instead
of numbers so I went with that.

The issue was that the player was doing 1-10 damage with its fist, and
we need to do one attack on the zombie but _NOT_ kill it. Doing 10
damage could kill the zombie. The test is making sure that the player,
with a 10 second resource wait time, doesn't have enough stamina to do
more than one attack before we can check that the first attack succeeded
(i.e. the zombie has less than its initial 10 hitpoints).

co-authored-by: Terry Wilcox <terry.wilcox@mac.com>
co-authored-by: Victor Savkov