Hack in a fix for integer regex
I have a regex to see if anything in the event is a PID (e.g. <0.20.0>)
so I can look up the ID that goes with that PID. Some of the
JSON-compatible values in the event are integers and calling
match(regex) on them breaks the page. I don't have any visible error
reporting so it just looks like the events stop as soon as a bug is
encountered.