Add experience gain
Characters can have an experience object that tracks both the experience
they give and the experience they've gained.

The experience object has a single rule that watches for its owner to be
killed and then gives the killer some experience through a {Char, gains,
Amount, experience} event.

I started out modifying the {Killer, killed, Victim, with, Attack} event
to add the exp, but objects don't need to see the two events together.

We may eventually have to add what caused the experience, but that can
be added to the experience gained event.