WIP: Achievements
We're going to demonstrate the flexibility of EGRE_MUD_1 by handling
triggers (quests, levels, achievements) in different ways:
- pre-complete quest/achievement
- post-complete quest/achievement
- visible/invisible quest/achievement

So far we're tracking past user actions with a metrics process and
achievements can check for existing completion metrics on startup using
the new init functionality: egre_object will automatically send each
object a {self(), init} event on startup.

Adding an achievement to a player on the fly required registering the
achievement with the player before the achievement could operate in the
graph; this is because until the player knows about the achievement
(i.e. has it's pid as a property), no events will make out to the
achievement; the achievement can send events in, through the player, but
then won't hear anything. So, before the achievement sends any messages
out, it must register with the player so that responses don't come back
before the achievement can possibly see them. That is, it's no sense
kicking off an event in the graph if there's no way for any events can
get back to you. Took me a while to figure this out.

See "quest_thoughts" in notes for more details.