Achievement ignores historical events
Added an "allow_previous" flag on achievements MUD creators can choose
whether players get credit for previous work when an achievement is
introduce. It's Oct 31 today, so a creator might want to add a Halloween
achievement for killing a halloween-appropriate monster X number of
times _today_. (Or whatever, doesn't have to be killing, could be
collecting candy or saying "Trick or Treat" some number of times.