Attack quests
Added attack quests that can rely on context.

Since the "effect" rules actually do the damage to a character, and can
potentially kill the character, they have to carry the context of the
attack containing anything relevant to any kill quest. In the case of
the "finish_quest" test, the quest needs to know what the attacker was
wearing, what time it was, and what the character was wielding. That
means that all wielded weapons, all worn clothing, and the time need to
be added to the context.

Aaaaand, I just realize that I'm not checking if the weapon is actually
_wielded_ and if the clothing is actually _worn_. You can't just _have_
a glove, you have to be wearing the glove.

I was thinking of making the context a map, but I don't need to pick out
pieces of the context by a key, I just need to make sure that the
context is a superset of the required context. Key-value pairs works
fine for this.