Use map instead of record for first param
The attempt function takes a convenience parameter as the first
parameter to make matching on certain fields easier. When I started the
MUD matching on maps (and even maps themselves) wasn't availble.

Props is likely no longer necessary since all properties are in the new
map.

Also fixed some bugs.

The decompose test was setting an mud app env var and screwing up the
get_experience_from_killing test, so we've set a reasonable value in the
init_testcase function.

I had to add a type to the fist so that the kill rats ... test could add
that the the context.