Differentiate between "having" and "wearing"
For "wearing" or "wielding" quest objectives, you have to have the item
of clothing on a "wearing" body part and you have to have the weapon on
a "wielding" body part. A sword in the hand might be considered
"wielded", where a sword in a sheath might be considered "worn"; the
same goes for a glove in a pocket or a glove on the hand. (I'm not sure
if we can distinguish yet between a glove _in_ a hand or a glove _on_ a
hand. I don't think all MUDs will need to make that distinction.)

Update "complete_quest" to attack 1 rat without the glove on, then put
the glove on and attack a 2nd rat; at this point the quest should not be
complete because the glove is required to be worn for both rats. After
that, kill a third rat (the second while wearing the glove) and make
sure the quest is now complete.

Use macros for creating parts of test worlds. I keep forgetting to add
rules to objects, and I keep having to do boilerplate. Using macros
makes this less error prone. DRY.

Update notes. I ended up making it so that a character has to ask its
life object if it can inject itself; if not, then the life adds the
character to the "dead" list. This means characters have to handle the
first message where their name is found, or the subsequent messages with
the "dead" list in them.

Start adding {rules_module, ...} properties to success logging so I can
see which rules module responded: *all* rules modules for a subscribed
object are called when an attempt succeeds. Typically, I'll keep the
success function that corresponds to a subscribing attempt function in
the same module.