Only wielded weapons add "wielded ____" context
When a character affects another character while wielding a weapon, that
weapon will add {wielding, WeaponType} to the context of the effect.
This context is transfered to the kill event if this effect kills the
target. Quests, upon seeing the kill event, will check the context to
see if the player they're attached to has completed all of the quest
objectives, e.g. [{kill, rat}, {wielding, fist}].

This fixes a bug where the weapon was added to the context as {wielded,
WeaponType} even though the weapon might not have been attached to a
body part that is considered a "wielding" body part. Weapons should have
a property called "wielding_body_parts" with a list of body parts that
would be considered to be wielding the weapon if the weapon is on that
body part.