Expand quest capabilities
We intend to suport the following:
- having multiple active quests
- requesting a quest from a quest giver by name
- turning in multiple quests
- asking for a listing of quests
  - available
  - active
- not listing quests from a quest giver that the player is already
  working on
- quest rewards:
  - attribute modifications (e.g. +1 to charisma)
  - items
  - money

In order to accomplish that we broke the quest rules up into multiple
different rules modules:
- quest_ack: join a created quest to its player
- quest_killed: check objectives when something is killed
- quest_turn_in: turn in quests and get rewards
- quest_get: figure out what quests are already taken
- char_quest_giver: listing, giving out and turning in quests
- char_quest_doer: connecting a newly created quest to the player

So far, quests don't need their own specific rules modules.

The "turn in" test tests that a quest reward that updates an attribute
works as intended.

There's a love triangle between players, quests, and quest givers:

quest giver <---- quest <----> player

The quest knows who it belongs to (e.g. player) and where it came from
(e.g. shopkeeper quest giver). The player knows about the quest, but the
quest giver doesn't. The player object is responsible for tracking its
own quests: we don't expect the quest giver to track all the quests it
gives out for each player.

In order for a quest giver to give a listing of quests available, or to
turn in quests, the quest giver has to send out an event to poll for all
the quests for a given player created by that quest giver.

We're not sure what to do yet with completed quests: do we record them
all in one process?

Co-authored-by: Terry Wilcox <terry.wilcox@mac.com>