Add a DB logging view in the test runner
In order to see which pids belong to which IDs, the log table has to be
joined to the pid_id table a bunch of times. You can't drop a table
without dropping views of that table, so it's just easiest to recreate
the view in the test runner after the tables have for sure been
(re-)created.