Improve DB logging
Make the DB logging fields smaller to fit more on the screen and make
things easier to see at a glance.

Add a macro to automatically add the current module in succeed log
properties with the "rules_" prefix removed.

Since every rules module is run on a subscribed event, any rules module
that runs needs report it's module name in the logs. For attempts, only
one rules module can handle the attempt, so we can log which one it was
automatically. Many rules modules won't have anything to do with a
success, so they don't log anything.