Quest listings
When you say "quests", any NPC with quests should list out what quests
they have that you don't already have.

Still need a test for trying to get a quest that the player has already
turned in.

Quests have three stages so far:

| complete? | turned in? |
|           |            |  - Not complete, not turned in
|     X     |            |  - Complete, but not turned in
|     X     |     X      |  - Complete and turned in

Might want to allow quests that are turned in to be done again.