New #result{} record
There are two reasons why I'm switching from a free-form tuple as the
attempt/1 result to a record:

1) In EGRE I have three functions to ensure that all the necessary result
fields are filled in, and in the proper order:
    ensure_message,
    ensure_context,
    ensure_log.

It would be much easier if I could just check record fields instead of
figuring out what was returned using guard functions with type checks.

2) I want to add fields like "Context" to the result, but I don't want
to have to mess with all the different rules modules. This way I can
just add fields--if I need any more--without updating the code.
(At least, I think that was the second reason.)

I thought about using a map, but I want defaults; if I used a map then
I'd need to either merge an existing map, or have a function that
creates the map with details. Having an existing map with defaults that
I could merge with in EGRE would be fine, but this is what records are
for: a known set of fields with known types and defaults.

I used sed to make some of the changes and I haven't run tests yet, so
everything might be borked.

e.g. sed -i -e \
     's/{succeed, _Subscribe = false, Props, Log}/?SUCCEED_NOSUB/g' \
     src/rules/rules*.erl

I decided to make some macros to handle the common result cases because
I was repeating myself over and over. The macros require the property
and log variables to be "Props" and "Log", but I feel like that's a
small price to pay; also, Log is required with the macros, but I want
Log to be standard anyway. I removed any log/1 calls and replaced them
with returning the log proplist with the #return{} record.

There was one module that wasn't returning a value for "should
subscribe"; I have no idea how I got away with that, or if there was a
bug lurking in there. A test might be needed. I think it was
rules_subitem_attack.erl.

Hmm, this record should actually be defined in EGRE, as EGRE will use it
in the code. I don't think I've updated EGRE to handle it. I was in the
middle of threading the new Context return value through everything when
I decided to switch to the record.