Convenience compile script
Since we have code spread across three projects and we're working on
them all at once, it's nice to pull the dependencies in directly and
compile them altogether.

Previously I was compiling the code whenever I ran the tests, but that
makes it hard to see warnings, and it's a waste of time when there's a
lot of parse transform and compile errors to work through.