Context in events is arduous
Move context of events that aren't directly relevant to the event out
into the Context return field in the #result{} record. This way, we can
clutter up results with context about the state of the world at the time
of the event without affecting the protocol.

Add new quest rule and test for "get item" quests using the new Context
 #result{} field.

Co-authored-by: Terry Wilcox <terry.wilcox@mac.com>