Forgot to remove old regex
Since I'm logging PIDs as just the middle number I need to change the
regex I'm using to find PIDs in messages.

Unfortunately, the regex is picking up non-PID numbers. D'oh!

I might need to log the PIDS with markers, like {123}. The point was to
save space because the rest of the PID is useless. There are five or six
different possible PIDs in each log message and they take up space on
the screen when doing DB queries.