Fix player_attack test.
Use 'infinity' for "numer of times" a resource should be reserved for.
(Times refers to the number of allocations, not the number of ticks.)

I broke this after adding the ability to reserve a resource for a single
allocation to cast a spell.

Standardize on resource reservations being "for" a purpose, not "to" a
purpose. We could go either way on this, it just needs to be consistent.
It might be nice to have functions that compose events that are used in
multiple places.

Also, don't crash when a character doesn't have a name.