Don't damage dead characters
If two attack effects are generated in quick succession and the first
one kills the target, the second one won't be stopped by the target
being dead. Instead, if a target's HP is already at or below zero then
we just ignore any more damage.

I had an issue where a character was getting double exp because the
first effect killed the target character, and a "killed" message was
generated, and then a second effect further damaged the target
character, triggering a second "killed" message. The attacker got exp
for killing the target twice.