Convenience functions for tracing and profiling
I also fixed a test (if a wait_loop finds the expected value of false,
then it returns true; i.e. "the expected value was found")

I was messing around with the attack_with_modifiers test because one of
the attack hit/effect roll events wasn't formatted correctly and it was
generated a LOT of extra function calls.

So, now there's functions to add recon_tracing and write it to file, and
cprof, and write it to file. The files end up in the CT test directory
under logs/ct_run......./{recon,cprof}.