Characters can buy and sell items
Here's a rough idea of the protocol:

1. Player types "buy <ItemName>" -> event: {<Player> buy <ItemName>}
2. Item injects name -> event: {<Player> buy <Item>}
3. Seller has item for sale ->
    event: {<Player> buy <Item>
            from <Seller>
            for <Cost>}
4. Player creates escrow process,
    event: {<Player> buy <Item>
            from <Seller>
            for <Cost>
            with <Escrow>}
5. Escrow event: {<Player> reserve <Cost> for <Escrow>}
6. Escrow event: {<Seller> reserve <Item> for <Escrow>}
7. Escrow event: {<Item> reserve self for <Escrow>}
8. Player reserves money on attempt
9. Seller reserves item on attempt
10. Item reserves self on attempt
11. Escrow event: {<Player> spend <Cost> because <Escrow>}
12. Escrow event: {<Seller> accrue <Cost> because <Escrow>}
13. Escrow event: {<Item> move from <Seller> to <Buyer> because <Escrow>}
14. Item: {<Item> move from <Seller> to <Buyer>}

The buyer, seller, and item need to update their properties when they
see the reserve attempt because if they allow the attempt to succeed
because they requisite state to reserve, that state might change before
the event succeeds: e.g. if the player character has enough money to buy
the item when the {<Player>, reserve, <Cost>, for, <Escrow>} event is
attempted, then the player needs to reserve that money before any other
event is handled.

If any reserve event fails the escrow transaction is canceled and
rolled back.

ToDo:
- Tell the player if the item doesn't exist if no item injects its own
name
- Tell the player if the item isn't for sale because no seller resent
the buy message with "from <Seller> for <Cost>"
- Tell the player if the escrow transaction succeeds
- Tell the player if the escrow transaction fails