User notified when item not for sale
The item listens for the unexpected success of {Char, buy, Item}; the
reason the success is unexpected is because if the item is for sale then
the seller should resend the event as {Char, buy, Item, from, Seller,
for, Cost}; this means the original {Char, buy, Item} event simply
disappears and the item itself, although subscribed, never sees a
succeed event. If the {Char, buy, Item} event succeeds, the item wasn't
for sale, but did exist, so we notify the player that the item is not
for sale.

Co-authored-by: Terry Wilcox <terry.wilcox@mac.com>