Tell user when they have insufficient funds
There are two places in the code, in the buy transaction, where the
transaction is canceled if the user doesn't have the funds: one is when
the {Char, buy, Item, from, Seller, for, Cost} attempt gets to the
buying character and the buying character doesn't have enough money; two
is when the user *had* money when the {Char, buy, ...} attempt happened,
but then didn't have the money once the escrow {Buyer, reserve, Cost,
for, Escrow} event happened.

In order to test the second event the buyer character had to have money
when the first attempt happened ({Char, buy, ...}) but then had to _not_
have the money for the second event. To do this we added the ability in
EGRE to use a map of functions as a rules module instead of a module.
You pass in a map with:
  attempt => {FauxModuleName, AttemptFun},
  succeed => SucceedFun,
  fail => FailFun.

The reason attempt has a pretend module name is because the attempt code
logs which rules module returned something other than undefined for the
attempt.

Co-authored-by: Terry Wilcox <terry.wilcox@mac.com>