Failure to reserve cancels escrow
If the buyer gets past the initial money check and the escrow is
started, but then the buyer fails to reserve the money, then the escrow
should unreserve any other successful reservations. In this case the
other two legs of the escrow were listed as succesful (because I set the
world up that way from the start), so they have to be unreserved.

If I don't make the escrow wait for the success of the unreserve message
({unreserve, for, self()}), then the unreserve messages don't seem to
make it past the attempt stage to the succeed stage: the seller and the
item see the message, and even subscribe to it, but their succeed/1
functions are never called. I thought maybe I could just prevent the
escrow process from subscribing to the unreserve event, but that didn't
work. (Maybe I did it wrong.) So, for now, I'm making the escrow process
wait until it sees the success of the unreserve event before shutting
down.

Since each party in the escrow transaction listens for the same
unreserve message, I only need to send the message once, as opposed to
sending the message to each object involved in the transaction. At first
I was getting each party to the escrow transaction that had previously
succeeded and sending that process its own unreserve event. If I do
that, then each process might get the message twice (i.e. if only one
party fails to reserve then only two parties can have successfully
reserved, kicking off two unreserve events.)

I had to add a new clause in EGRE for world properties that have a key
pointing to a two-tuple of pids: {Key, {Pid1, Pid2}}. Since this case is
more specific than {Key, {Pid, Value}}, it has to come first, or else
the second {pid, Value} tuple is left as is.