We only need top level id & pid props
When logging, EGRE no longer looks up the ID and icon of every property
of the logging object that is a pid.

e.g. if Object A has properties [{room, Pid1} | _], then EGRE won't look
up the ID and icon for Pid1 and add it to the list of properties as
{room_id, RoomId}, {room_icon, RoomIcon}.

The logging JS will now track all the top-level IDs and PIDs in the
"procIds" map, and use that for all PID lookups.

Also, PIDs are now logged by EGRE as "|123|": i.e. the middle number of
the PID (which assumes <0.XXX.0>) surrounded by pipe symbols. This
allows us to only look up IDs for PIDs instead of any integer.

Previously I was using 'process_id' and 'process' to look up the ID and
PID of the logging object, but now I'm using id and pid. 'process' is
added by egre_event_log, 'pid' and 'id' are added by egre_object when
the object starts. I can probably do away with 'process' altogether.