Fix log database id column references
The id column of the logging object is now 'id', instead of 'pid_id'.
EGRE used to lookup all properties of an object with a PID value and
create a key called "<key>_id", which pointed to the ID of that PID.
e.g. {pid, <0.1.0>} causes a new property called {pid_id, room}.

However, I took out the in-depth property ID lookups and we're using the
'id' property that egre_object puts in the objects properties.