"Event" instead of "Msg"
I'm slowly transitioning to calling the pieces of data that flow through
the graph and trigger new pieces of data "events". This is distinct from
Erlang messages. Events are transmitted as messages, and more
specifically as tuples that have a distinct shape signalling which exact
event they are. Events in the graph need only be unique from other,
unrelated events.